+++ app/src/main/java/tn/amin/keyboard_gpt/SPManager.java
@@ -12,8 +12,9 @@ import java.util.stream.Collectors;
 
 import tn.amin.keyboard_gpt.instruction.command.Commands;
 import tn.amin.keyboard_gpt.instruction.command.GenerativeAICommand;
-import tn.amin.keyboard_gpt.llm.LanguageModel;
+import tn.amin.keyboard_gpt.llm.client.LanguageModel;
 import tn.amin.keyboard_gpt.listener.ConfigInfoProvider;
+import tn.amin.keyboard_gpt.ui.UiInteractor;
 
 public class SPManager implements ConfigInfoProvider {
     protected static final String PREF_NAME = "keyboard_gpt";
