+++ app/src/main/java/tn/amin/keyboard_gpt/instruction/command/AbstractCommand.java
@@ -1,8 +1,5 @@
 package tn.amin.keyboard_gpt.instruction.command;
 
-import tn.amin.keyboard_gpt.GenerativeAIController;
-import tn.amin.keyboard_gpt.UiInteractor;
-
 public abstract class AbstractCommand {
     abstract public String getCommandPrefix();
 }
