+++ app/src/main/java/tn/amin/keyboard_gpt/instruction/command/CommandManager.java
@@ -5,9 +5,9 @@ import java.util.List;
 import java.util.Map;
 
 import tn.amin.keyboard_gpt.SPManager;
-import tn.amin.keyboard_gpt.UiInteractor;
+import tn.amin.keyboard_gpt.ui.UiInteractor;
 import tn.amin.keyboard_gpt.listener.ConfigChangeListener;
-import tn.amin.keyboard_gpt.llm.LanguageModel;
+import tn.amin.keyboard_gpt.llm.client.LanguageModel;
 
 public class CommandManager implements ConfigChangeListener {
     private final static Map<String, AbstractCommand> STATIC_COMMAND_MAP = Map.of(
