+++ app/src/main/java/tn/amin/keyboard_gpt/listener/ConfigInfoProvider.java
@@ -2,7 +2,7 @@ package tn.amin.keyboard_gpt.listener;
 
 import android.os.Bundle;
 
-import tn.amin.keyboard_gpt.llm.LanguageModel;
+import tn.amin.keyboard_gpt.llm.client.LanguageModel;
 
 public interface ConfigInfoProvider {
     LanguageModel getLanguageModel();
