+++ app/src/main/java/tn/amin/keyboard_gpt/llm/client/GroqClient.java
@@ -1,4 +1,4 @@
-package tn.amin.keyboard_gpt.llm;
+package tn.amin.keyboard_gpt.llm.client;
 
 import org.json.JSONArray;
 import org.json.JSONException;
@@ -16,10 +16,13 @@ import java.util.stream.Collectors;
 import tn.amin.keyboard_gpt.MainHook;
 import tn.amin.keyboard_gpt.llm.publisher.ExceptionPublisher;
 import tn.amin.keyboard_gpt.llm.publisher.InputStreamPublisher;
-import tn.amin.keyboard_gpt.llm.publisher.SimpleStringPublisher;
 
-public class ChatGPTClient extends LanguageModelClient {
+public class GroqClient extends ChatGPTClient {
     @Override
+    public LanguageModel getLanguageModel() {
+        return LanguageModel.Groq;
+    }
+
     public Publisher<String> submitPrompt(String prompt, String systemMessage) {
         if (getApiKey() == null || getApiKey().isEmpty()) {
             return LanguageModelClient.MISSING_API_KEY_PUBLISHER;
@@ -47,7 +50,7 @@ public class ChatGPTClient extends LanguageModelClient {
             JSONObject rootJson = new JSONObject();
             rootJson.put("model", getSubModel());
             rootJson.put("messages", messagesJson);
-            rootJson.put("stream", false);
+            rootJson.put("stream", true);
 
             con.setDoOutput(true);
 
@@ -60,31 +63,31 @@ public class ChatGPTClient extends LanguageModelClient {
             MainHook.log("Received response with code " + responseCode);
 
             if (responseCode == 200) {
-                BufferedReader reader = new BufferedReader(new InputStreamReader(con.getInputStream()));
-                String response = reader.lines().collect(Collectors.joining(""));
-                JSONObject responseJson = new JSONObject(response);
-                if (responseJson.has("choices")) {
-                    JSONArray choices = responseJson.getJSONArray("choices");
-                    for (int i = 0; i < choices.length(); i++) {
-                        JSONObject choice = choices.getJSONObject(i);
-                        if (choice.has("role") && "assistant".equals(choice.getString("role"))) {
-                            return new SimpleStringPublisher(choice
-                                    .getJSONObject("message")
-                                    .getString("content"));
+                return new InputStreamPublisher(con.getInputStream(), line -> {
+                    line = line.trim();
+//                    MainHook.log("line is " + line);
+                    if (line.isEmpty() || line.endsWith("[DONE]")) {
+                        return "";
+                    }
+                    if (line.startsWith("data:")) {
+                        line = line.substring("data:".length()).trim();
+                    }
+                    try {
+                        JSONObject choice = new JSONObject(line)
+                                .getJSONArray("choices")
+                                .getJSONObject(0);
+                        if (choice.has("delta")) {
+                            return extractContent(choice.getJSONObject("delta"));
                         }
+                        return extractContent(choice.getJSONObject("message"));
+                    } catch (JSONException e) {
+                        MainHook.log(e);
+                        return "";
                     }
-                    if (choices.length() > 0) {
-                        return new SimpleStringPublisher(choices.getJSONObject(0)
-                                .getJSONObject("message")
-                                .getString("content"));
-                    }
-                    else {
-                        throw new JSONException("choices has length 0");
-                    }
-                } else {
-                    throw new JSONException("no \"choices\" attribute found");
-                }
-            } else {
+                });
+            }
+
+            else {
                 BufferedReader reader = new BufferedReader(new InputStreamReader(con.getErrorStream()));
                 String response = reader.lines().collect(Collectors.joining(""));
                 JSONObject responseJson = new JSONObject(response);
@@ -92,10 +95,7 @@ public class ChatGPTClient extends LanguageModelClient {
                     JSONObject errorJson = responseJson
                             .getJSONObject("error");
                     String message = errorJson.getString("message");
-                    String type = "";
-                    if (errorJson.has("type")) {
-                        type = errorJson.getString("type");
-                    }
+                    String type = errorJson.getString("type");
 
                     throw new IllegalArgumentException("(" + type + ") " + message);
                 }
@@ -108,8 +108,12 @@ public class ChatGPTClient extends LanguageModelClient {
         }
     }
 
-    @Override
-    public LanguageModel getLanguageModel() {
-        return LanguageModel.ChatGPT;
+    private String extractContent(JSONObject message) throws JSONException {
+        if (!message.has("content")) {
+            return "";
+        }
+
+        return message.getString("content");
     }
+
 }
