+++ app/src/main/java/tn/amin/keyboard_gpt/llm/client/HuggingChatClient.java
@@ -1,4 +1,4 @@
-package tn.amin.keyboard_gpt.llm;
+package tn.amin.keyboard_gpt.llm.client;
 
 import org.json.JSONException;
 import org.json.JSONObject;
