+++ app/src/main/java/tn/amin/keyboard_gpt/llm/client/OpenRouterClient.java
@@ -1,4 +1,4 @@
-package tn.amin.keyboard_gpt.llm;
+package tn.amin.keyboard_gpt.llm.client;
 
 public class OpenRouterClient extends ChatGPTClient {
     @Override
